Finish Dashboard project with theme
Dashboard is not fully functional. However, it finishes the assignment
given by TOP (The Odin Project). A directory for the Roboto font family
exists, but isn't used internally. Theme is centered around a fictional
science group in a game called Blue Archive. All characters are owned by
Nexon and I do not claim any original creation of the characters or
works used.